+++ api/src/main/kotlin/github/com/hukuta94/delivery/api/startup/configuration/adapter/orm/OrmCourierRepositoryConfiguration.kt
@@ -1,6 +1,6 @@
 package github.com.hukuta94.delivery.api.startup.configuration.adapter.orm
 
-import github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.CourierJpaRepository
+import github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.jpa.CourierJpaRepository
 import github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.OrmCourierRepositoryAdapter
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
