+++ api/src/main/kotlin/github/com/hukuta94/delivery/api/startup/configuration/adapter/orm/OrmOrderRepositoryConfiguration.kt
@@ -1,6 +1,6 @@
 package github.com.hukuta94.delivery.api.startup.configuration.adapter.orm
 
-import github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.OrderJpaRepository
+import github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.jpa.OrderJpaRepository
 import github.com.hukuta94.delivery.infrastructure.adapter.orm.repository.OrmOrderRepositoryAdapter
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
